wiederkehrende Einnahmen und Ausgaben können erfasst werden neue Funktion in der Navigation Bar um die Daten löschen zu können (ersetzt das manuelle ausführen der zwei Zeilen im Code)

+++ app/src/main/java/com/example/finanzmanager/Objects/Month.java
@@ -4,9 +4,9 @@ public class Month {
 
     private int year;
     private int month;
-    private int sum_income;
-    private int sum_expense;
-    private int balance;
+    private double sum_income;
+    private double sum_expense;
+    private double balance;
 
     Month(int year, int month) {
         this.year = year;
@@ -16,18 +16,18 @@ public class Month {
         balance = 0;
     }
 
-    public int getSumIncome() {return sum_income;}
-    public int getSumExpense() {return sum_expense;}
-    public int getBalance() {return balance;}
+    public double getSumIncome() {return sum_income;}
+    public double getSumExpense() {return sum_expense;}
+    public double getBalance() {return balance;}
     public int getYear() {return year;}
     public int getMonth() {return month;}
 
-    void setIncome(int value) {
+    void setIncome(double value) {
         sum_income = sum_income + value;
         balance = sum_income - sum_expense;
     }
 
-    void setExpense(int value) {
+    void setExpense(double value) {
         sum_expense = sum_expense + value;
         balance = sum_income - sum_expense;
     }
